chore(virtctl): Move start command to own file
