+++ pkg/virtctl/vm/start.go
@@ -1,0 +1,82 @@
+/*
+ * This file is part of the KubeVirt project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Copyright 2023 Red Hat, Inc.
+ *
+ */
+
+package vm
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/spf13/cobra"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/tools/clientcmd"
+	v1 "kubevirt.io/api/core/v1"
+
+	"kubevirt.io/kubevirt/pkg/virtctl/templates"
+)
+
+const (
+	COMMAND_START = "start"
+	pausedArg     = "paused"
+)
+
+var (
+	startPaused bool
+)
+
+func NewStartCommand(clientConfig clientcmd.ClientConfig) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:     "start (VM)",
+		Short:   "Start a virtual machine.",
+		Example: usage(COMMAND_START),
+		Args:    templates.ExactArgs("start", 1),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			c := Command{command: COMMAND_START, clientConfig: clientConfig}
+			return c.startRun(args)
+		},
+	}
+	cmd.Flags().BoolVar(&startPaused, pausedArg, false, "--paused=false: If set to true, start virtual machine in paused state")
+	cmd.Flags().BoolVar(&dryRun, dryRunArg, false, dryRunCommandUsage)
+	cmd.SetUsageTemplate(templates.UsageTemplate())
+	return cmd
+}
+
+func (o *Command) startRun(args []string) error {
+	var dryRunOption []string
+	vmiName := args[0]
+
+	virtClient, namespace, err := GetNamespaceAndClient(o.clientConfig)
+	if err != nil {
+		return err
+	}
+
+	if dryRun {
+		dryRunOption = []string{metav1.DryRunAll}
+		fmt.Printf("Dry Run execution\n")
+	}
+
+	err = virtClient.VirtualMachine(namespace).Start(context.Background(), vmiName, &v1.StartOptions{Paused: startPaused, DryRun: dryRunOption})
+	if err != nil {
+		return fmt.Errorf("Error starting VirtualMachine %v", err)
+	}
+
+	fmt.Printf("VM %s was scheduled to %s\n", vmiName, o.command)
+
+	return nil
+}
